pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fb-deal-alert"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "sentence-transformers>=2.7",
    "openai>=1.30",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "loguru>=0.7",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
markers = ["integration: requires network/model download"]
